// src/app/lib/enums/product.enum.ts
// src/enums/product.enum.ts

export enum ProductCategory {
  BEDROOM = 'BEDROOM',
  OFFICE = 'OFFICE',
  KITCHEN = 'KITCHEN',
  OUTDOOR = 'OUTDOOR',
  OTHER= 'OTHER',
  KIDS = 'KIDS',
}

export enum ProductMaterialType {
  WOOD = 'WOOD',
  METAL = 'METAL',
  GLASS = 'GLASS',
  PLASTIC = 'PLASTIC',
  FABRIC = 'FABRIC',
  LEATHER = 'LEATHER',
  MARBLE = 'MARBLE',
  BAMBOO = 'BAMBOO',
}

export enum ProductStyleType {
  MODERN = 'MODERN',
  CONTEMPORARY = 'CONTEMPORARY',
  TRADITIONAL = 'TRADITIONAL',
  RUSTIC = 'RUSTIC',
  INDUSTRIAL = 'INDUSTRIAL',
  SCANDINAVIAN = 'SCANDINAVIAN',
  BOHEMIAN = 'BOHEMIAN',
  VINTAGE = 'VINTAGE',
  MINIMALIST = 'MINIMALIST',
}
export  enum ProductStatus{
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}